Further consolidate completion fixtures
- Merge 4 inheritance fixtures into single InheritanceCompletion.php
- Remove NullsafeUser/NullsafeUsage from MethodAccess.php, use User
