--- tests/Fixtures/src/Completion/ParentAccess.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Fixtures\Completion;
-
-use Fixtures\Inheritance\ParentClass;
-
-class ParentAccess extends ParentClass
-{
-    public function __construct(string $name)
-    {
-        parent::/*|parent_access*/
-    }
-}
